feat: Add env last seen to toggle overview
