+++ frontend/src/component/feature/FeatureView/FeatureEnvironmentSeen/useLastSeenColors.ts
