+++ frontend/src/component/feature/FeatureView/FeatureOverview/FeatureOverviewSidePanel/FeatureOverviewSidePanelDetails/FeatureOverviewSidePanelDetails.tsx
@@ -3,6 +3,8 @@ import { styled } from '@mui/material';
 import { useLocationSettings } from 'hooks/useLocationSettings';
 import { formatDateYMD } from 'utils/formatDate';
 import { parseISO } from 'date-fns';
+import { FeatureEnvironmentSeenCell } from '../../../../../common/Table/cells/FeatureSeenCell/FeatureEnvironmentSeenCell';
+import { FeatureEnvironmentSeen } from '../../../FeatureEnvironmentSeen/FeatureEnvironmentSeen';
 
 const StyledContainer = styled('div')(({ theme }) => ({
     display: 'flex',
@@ -21,6 +23,16 @@ interface IFeatureOverviewSidePanelDetailsProps {
     header: React.ReactNode;
 }
 
+const FlexRow = styled('div')({
+    display: 'flex',
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+});
+
+const StyledDetail = styled('div')({
+    justifyContent: 'center',
+});
+
 export const FeatureOverviewSidePanelDetails = ({
     feature,
     header,
@@ -30,15 +42,22 @@ export const FeatureOverviewSidePanelDetails = ({
     return (
         <StyledContainer>
             {header}
-            <div data-loading>
-                <StyledLabel>Created at:</StyledLabel>
-                <span>
-                    {formatDateYMD(
-                        parseISO(feature.createdAt),
-                        locationSettings.locale
-                    )}
-                </span>
-            </div>
+            <FlexRow>
+                <StyledDetail>
+                    <StyledLabel>Created at:</StyledLabel>
+                    <span>
+                        {formatDateYMD(
+                            parseISO(feature.createdAt),
+                            locationSettings.locale
+                        )}
+                    </span>
+                </StyledDetail>
+                <FeatureEnvironmentSeen
+                    featureLastSeen={feature.lastSeenAt}
+                    environments={feature.environments}
+                    sx={{ pt: 0 }}
+                />
+            </FlexRow>
         </StyledContainer>
     );
 };
